Move not-verified message box to local navigation
