+++ src/client/components/CompanyTabbedLocalNavigation/index.jsx
@@ -3,6 +3,13 @@ import PropTypes from 'prop-types'
 
 import CompanyLocalTab from './CompanyLocalTab'
 import styled from 'styled-components'
+import Paragraph from '@govuk-react/paragraph'
+import WarningText from '@govuk-react/warning-text'
+import Details from '@govuk-react/details'
+import Button from '@govuk-react/button'
+import urls from '../../../lib/urls'
+import StatusMessage from '../../../client/components/StatusMessage'
+import { BLACK } from '../../../client/utils/colours'
 
 const StyledGridRow = styled.div`
   margin-right: -15px;
@@ -41,11 +48,42 @@ const StyledUnorderedList = styled.ul`
   }
 `
 
+const StyledLink = styled('a')`
+  margin-bottom: 0;
+`
+
 const CompanyTabbedLocalNavigation = (props) => {
   const { localNavItems } = props
+  const company = props.company
+  const showMatchingPrompt =
+    !company.duns_number && !company.pending_dnb_investigation
 
   return (
     <StyledGridRow>
+      {showMatchingPrompt && (
+        <StatusMessage colour={BLACK} id="ga-company-details-matching-prompt">
+          <WarningText>
+            Business details on this company record have not been verified and
+            could be wrong.
+          </WarningText>
+          <Details summary="Why verify?">
+            <Paragraph>
+              Using verified business details from a trusted third-party
+              supplier means we can keep certain information up to date
+              automatically. This helps reduce duplicate records, provide a
+              shared view of complex companies and make it more likely we can
+              link other data sources together.
+            </Paragraph>
+            <Paragraph>
+              Verification can often be done in just 4 clicks.
+            </Paragraph>
+          </Details>
+          <Button as={StyledLink} href={urls.companies.match.index(company.id)}>
+            Verify business details
+          </Button>
+        </StatusMessage>
+      )}
+
       <StyledGridColumn>
         <StyledNav aria-label="local navigation" data-test="tabbedLocalNav">
           <StyledUnorderedList data-test="tabbedLocalNavList">
